feat(cherryMarkdown): 去除markdown-it
