+++ package.json
@@ -64,8 +64,6 @@
     "@babel/runtime": "^7.24.7",
     "@cherry-markdown/cherry-markdown-dev": "^0.9.4-dev.202507230237.4848b82",
     "@popperjs/core": "^2.11.8",
-    "@types/markdown-it": "^13.0.7",
-    "@vscode/markdown-it-katex": "^1.1.1",
     "cherry-markdown": "^0.9.4",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
@@ -73,9 +71,7 @@
     "htm": "^3.1.1",
     "immer": "^10.1.1",
     "lodash-es": "^4.17.21",
-    "markdown-it": "^14.1.0",
     "markdown-it-fence": "^0.1.3",
-    "markdown-it-link-attributes": "^4.0.1",
     "nanostores": "^0.11.4",
     "omi": "^7.7.13",
     "omi-reactify": "0.0.10",
