+++ src/chat-message/_example/markdown.tsx
@@ -6,16 +6,6 @@ import mdContent from '../../chatbot/mock/testMarkdown.md?raw';
 
 export default class BasicExample extends Component {
   render() {
-    return (
-      <t-chat-md-content
-        content={mdContent}
-        pluginConfig={[
-          {
-            preset: 'code',
-            enabled: true,
-          },
-        ]}
-      ></t-chat-md-content>
-    );
+    return <t-chat-md-content content={mdContent}></t-chat-md-content>;
   }
 }
