+++ src/chat-message/chat-item.tsx
@@ -1,5 +1,4 @@
 import './content/markdown-content';
-import './content/cherry-markdown-content';
 import '../collapse';
 import '../chat-loading';
 import '../attachments';
@@ -243,12 +242,12 @@ export default class ChatItem extends Component<ChatMessageProps> {
         if (isTextContent(content) || isMarkdownContent(content)) {
           // 正文回答
           return (
-            <t-chat-cherry-md-content
+            <t-chat-md-content
               key={elementKey}
               className={`${className}__detail`}
               {...chatContentProps?.markdown}
               content={content.data}
-            ></t-chat-cherry-md-content>
+            ></t-chat-md-content>
           );
         }
         // 自定义渲染slot
